refactor(parts): move Parts domain into src/SheetMusic.Api/Parts
Moves all music part artifacts (controller, commands, queries, entities,
request/view models, errors, and the search index model) out of the
legacy artifact-type folders and into a single domain-oriented
src/SheetMusic.Api/Parts/ folder, per the domain-folder refactor plan
(#192). Test files mirror the same layout under Tests/Parts/.

No behavior change; namespaces updated to match new locations and
legacy files that still reference Parts types got the necessary using
directives.

Closes #197
