+++ src/SheetMusic.Api.Test/Tests/Parts/Models/PutPartModel.cs
@@ -1,4 +1,4 @@
-﻿namespace SheetMusic.Api.Test.Models;
+﻿namespace SheetMusic.Api.Test.Tests.Parts.Models;
 
 public class PutPartModel
 {
