+++ src/SheetMusic.Api.Test/Utility/FakerFactory.cs
@@ -1,5 +1,6 @@
 ﻿using Bogus;
 using SheetMusic.Api.Test.Models;
+using SheetMusic.Api.Test.Tests.Parts.Models;
 using System;
 
 namespace SheetMusic.Api.Test.Utility;
