+++ src/SheetMusic.Api/Database/DatabaseSeeder.cs
@@ -3,6 +3,7 @@ using System.Text;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using SheetMusic.Api.Database.Entities;
+using SheetMusic.Api.Parts.Entities;
 
 namespace SheetMusic.Api.Database;
