+++ src/SheetMusic.Api/Database/Entities/Musician.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
+using SheetMusic.Api.Parts.Entities;
 
 namespace SheetMusic.Api.Database.Entities;
