+++ src/SheetMusic.Api/Database/Entities/SheetMusicPart.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using SheetMusic.Api.Parts.Entities;
 
 namespace SheetMusic.Api.Database.Entities;
