+++ src/SheetMusic.Api/Parts/Commands/AddPart.cs
@@ -1,13 +1,13 @@
 ﻿using MediatR;
 using SheetMusic.Api.Database;
-using SheetMusic.Api.Database.Entities;
-using SheetMusic.Api.Errors;
+using SheetMusic.Api.Parts.Entities;
+using SheetMusic.Api.Parts.Errors;
 using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.CQRS.Command;
+namespace SheetMusic.Api.Parts.Commands;
 
 public class AddPart(string name, int sortOrder, bool indexable) : IRequest<MusicPart>
 {
