+++ src/SheetMusic.Api/Parts/Commands/DeletePart.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.CQRS.Command;
+namespace SheetMusic.Api.Parts.Commands;
 
 public class DeletePart(Guid partId) : IRequest
 {
