+++ src/SheetMusic.Api/Parts/Commands/RemoveAliasFromPart.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.CQRS.Command;
+namespace SheetMusic.Api.Parts.Commands;
 
 public class RemoveAliasFromPart(Guid partId, string aliasToRemove) : IRequest
 {
