+++ src/SheetMusic.Api/Parts/Entities/MusicPartAlias.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SheetMusic.Api.Database.Entities;
+namespace SheetMusic.Api.Parts.Entities;
 
 public class MusicPartAlias
 {
