+++ src/SheetMusic.Api/Parts/Errors/AliasAlreadyAddedError.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SheetMusic.Api.Errors;
+namespace SheetMusic.Api.Parts.Errors;
 
 public class AliasAlreadyAddedError(string alias, string partName) : Exception($"Alias '{alias}' already exists for part '{partName}'")
 {
